[FIX] repair/saas~16.4: save operations as note
In old repair model, there was an 'Operations' tab referencing the
'repair_fee' field.
This tab was allowing to bind service type products to the repair.

As a repair order is now linked with a sale order rather than directly
bound to an invoice to enforce scope separation, the RepairFee model and
the related repair_fee field on the repair order have been removed.
It is now expected for the user to add those services to the linked sale order.

However, some users that have been using the old repair model stil have
pending repair orders with not or partially invoiced operations that
disappear during the upgrade process.

To allow them retrieve these data, we append an operations table in the
internal_notes when necessary.

Prerequisites :
- Odoo (Enterprise or Community) database in version < 16.4
- Repair module installed
- Existence of a Repair Order record with a non-empty repair_fee value

Steps to reproduce :
- Update the database to version >=16.4

opw-3861648
opw-3864870

Community : odoo/odoo#106911
Enterprise : odoo/enterprise#36088

[FIX] repair/saas~16.4: pre-commit hook

Declaring the utf-8 encoding is no more necessary.

closes odoo/upgrade#5951

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>