[IMP] sign/saas~17.2: remove ignored field
The ignored field in the sign.request.item model was used to stop notifications
from being sent for a request that you did not want to sign at that moment.
However, if you have ignored a sign request item, you probably will not sign
that request item in the future. So it is easier to just cancel the sign
request. This commit does exactly that by removing the ignored field and
setting the ignored sign requests items and associated sign requests to
cancelled.

Enterprise PR: odoo/enterprise#55463

task-3675303

closes odoo/upgrade#5744

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>