[FIX] project/saas~16.2: Move query to correct version
Issue:
Upgrading a customer from version 16.2 to 17.0 resulted in tasks
disappearing. This occurred because a query setting project_id to
null based on a unused field (display_project_id) was present in
the script.

Fix:
This commit moves the problematic query to the correct version,
ensuring it doesn't affect tasks during the upgrade process.

Testing:
Upgrading from both 16.2 and 16.0 to 17.0 confirmed the fix works
as expected.

References:
- https://github.com/odoo/odoo/pull/115546

closes odoo/upgrade#5985

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>