[FIX] industry_fsm/saas~16.5.1.0 : update custom worksheet report
If customer has created copy of standard worksheet report in any version before
version saas-16.4, then while upgrading from version saas-16.4 to 17, this
studio report using the key 'industry_fsm.worksheet%' which hasn't been
updated to 'industry_fsm_report.worksheet%' according to the standard changes.
This mismatch is causing traceback errors when attempting to print the report.

To resolve this issue, we need to update the arch_db,key and the report_name to
align with the standard change.

closes odoo/upgrade#5992

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>