[FIX] analytic/17.0: cleanup analytic distributions
The mixin field 'analytic_distribution' is supposed to be a json object
but some values valeus are found to be empty strings or 0 which is no
longer accepted. The index on the field is also modified and renamed but
the old index is not removed.

see https://github.com/odoo/odoo/pull/147664

closes odoo/upgrade#5993

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>