[IMP] mail/saas~17.3: improve message seen/unread tracking
In the community counterpart of this PR, the
field is introduced to keep track of the user's unread messages.

Previously, the  field was used but it is incorrect
as read and seen messages are different concepts.

This PR assigns the value of the  field to the newly
created  one in order to keep the information
about the last read message.

see odoo/odoo#163091

closes odoo/upgrade#5994

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>