[IMP] website_forum/saas~17.3: rename post column in forum_tag_rel table
At the bottom of a forum post page, we've added a list of 5 related
posts. The selection, based on the number of similar tags, is obtained
using an SQL query.

While writing this query, we noticed that the column storing post ids in
the relational table between posts and tags was entitled 'forum_id'.
This logically refers to another model: the forum model.

To clear up any confusion, we renamed the column to 'forum_post_id'.

see odoo/odoo#163002
task-3859111

closes odoo/upgrade#6015

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>