[FIX] mail/saas~17.3: fix incorrect new message separator value
In [1], the `new_message_separator` field was introduced to keep track
of the user unread messages. The first unread message is the one after
the last one the user saw (`seen_message_id` field).

The upgrade script sets the value of this new field to the value of the
`seen_message_id` while it should actually be `seen_message_id + 1`.

This PR corrects this mistake.

Steps to reproduce it:
- create a db with mail installed
- check the value of the `seen_message_id` field on the
`discuss_channel_member` table
- proceed to the upgrade to master
- the `new_message_separator` field on the `discuss_channel_member`
table was set to the value of `seen_message_id` while it should have
been `seen_message_id + 1`.

[1]: https://github.com/odoo/upgrade/pull/5994

closes odoo/upgrade#6019

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>