[FIX] sale_subscription/0.0.0: rename xmlids
In version saas-16.2 the sale subscription report started to inherit
sale.report which came possible with fields 'website_id' and
'is_abandoned_cart' based on whether 'website_sale' or
'website_sale_dashboard' were also installed.
In version saas-16.3 both fields became part of 'website_sale'.
The order of installation of modules dictates the module on the xmlid
for inherited fields on the subscription report, so they can have
the 'sale_subscription' as module which is not correct, and leads to
issues when the website_sale module is uninstalled, because those
inherited fields are not removed, causing failures in upgrades.

closes odoo/upgrade#5085

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>