[IMP] spreadsheet_edition/saas~17.3: change chart commands
This commit changes the way chart title and dataSets are stored in the
CREATE_CHART and UPDATE_CHART commands. This is done to be able to add
properties to the chart title and dataSets (e.g. the color of a text, ...)
`verticalAxisPosition` is also removed from the definition, except for
the waterfall charts

Before this commit:
```js
{
  type: "CREATE_CHART",
  definition: {
    type: "line",
    title: "My chart",
    dataSets: ["A1:B2", "C1:D2"],
    verticalAxisPosition: "left",
    ...
  }
}
```

After this commit:
```js
{
  type: "CREATE_CHART",
  definition: {
    type: "line",
    title: {
      text: "My chart",
    },
    dataSets: [
      {
        dataRange: "A1:B2",
      },
      {
        dataRange: "C1:D2",
      },
    ],
    ...
  }
}
```

see https://github.com/odoo/odoo/pull/166538

Task-id 3792009

closes odoo/upgrade#6029

Related: odoo/odoo#166538
Related: odoo/enterprise#63090
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>