[FIX] account/saas~13.5: recalculate balance on move lines
In some dbs, the `balance` field on `account.move.line`s is not computed correctly.
We recalculate it in Postgres similarly to how it's done in standard Odoo:
https://github.com/odoo/odoo/blob/f2765d2cab5671a010404c36842bf1b4c4d6350b/addons/account/models/account_move.py#L3976-L3979

Leaving the balance incorrect prevents the upgrade from adding the constraint
"account_move_line_check_amount_currency_balance_sign".

Part-of: odoo/upgrade#6043