[FIX] website/17.0: archive deprecated redirects
After this
[commit](https://github.com/odoo/odoo/pull/165083/commits/adaa80a2216f4667af6213cc43a28bc8768bfd5b),
a validation error has been added in v17 where we can no longer
configure `/` in the `url_to` field for redirects. Consequently,
as per the validation error, archived the existing redirects that
used `/` in the `url_to` field before v17.

If these redirects where / is used in the url_to field are active,
the issue faced in the upgraded version will be as follows:
For example, if a redirect has been created where url_from is set to /shop
and url_to is set to /, navigating to the homepage will display the shop page.
Conversely, navigating to the shop page will change the endpoint to /,
but the page displayed will still be the shop page.

opw - 3890040, 3916523

closes odoo/upgrade#6048

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>