[FIX] account_asset_ndt/15.0.1.0: Unlink wrong Journal Items
Unlink the wrong move line linked with the asset
before re-computing the asset field ‘non_deductible_tax_val’
because during the upgrade when it’s recomputing the asset
field ‘non_deductible_tax_val’ at that time constraint is violated

"You cannot create an asset from lines containing credit
and debit on the account or with a null amount"

This constraint was introduced in version saas-12-5:
https://github.com/odoo-dev/enterprise/commit/6e0d855af162aacfbe80df9ef199d0c12d029154
but the trigger the recompute, was introduced in version 15.0
https://github.com/odoo/enterprise/pull/23761/files#diff-f997fd2b42170395940ad5abbd161d6d888fdcec3a402008de9c028023fd6d9eR162
The problem is that we are recomputing a field here
that will also trigger the check.
Those assets are already wrong since long ago.
This patch is unlink the wrong move line linked with the asset since long ago.

upg - 1577976
[TGB](https://upgrade.odoo.com/web#id=696&cids=1&menu_id=107&action=178&model=upgrade.request.traceback.group&view_type=form)

closes odoo/upgrade#6060

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>