[IMP] base/0.0.0: remove mixin selections for unloaded models
When a module is partially uninstalled some of its models and fields
are not removed, including mixin selections. The mixin selections have
an xmlid that belongs to the mixin module, e.g. 'mail' so these records
end up being removed in the end of the migration because they were not
loaded. The removal of seelction fields via ORM needs the model on which
is is to be loaded, leading to the failure of the upgrade.

Here we detect the models that are uninstalled, and remove any
selection value that is generated from an installed module.

closes odoo/upgrade#6092

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>