[FIX] spreadsheet_edition/saas~17.3: fix chart commands migration
This commit fixes the commit 46085e8.
In the `UPDATE_CHART` command, some keys can be missing as this command
could historically take a diff instead of the full definition.

So, with this commit:
- Presence of key `title` was already checked -> no change
- Presence of key `dataSets` is now checked, and it's not necessary
anymore to check the type of the chart, all the charts with a key
`dataSets` should be transformed anyway.
- Removing of `verticalAxisPosition` would be better, but as we do not
have the opportunity to check that the type of the chart is `waterfall`
or not, it's better to keep this key. It's not a problem as the code
handle it.

This issue has been raised by NSE during the migration of odoo.com

Part-of: odoo/upgrade#6130
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>