[IMP] l10n_ch/saas~17.5: remove deleted invoice pdf views and report
Removes deleted views and records in the l10n_ch module. These views and
records have been removed to simplify and improve the performance of
generating PDF invoices for Switzerland. This reduces the number of calls
made to the function _run_wkhtmltopdf(), from 3 calls to 2. This is achieved
by adding an empty page, which still contains the header, to the generic
invoice template. Then the Swiss QR code, if it exists, is merged onto this
last added page.

See https://github.com/odoo/odoo/pull/168666
task-3964501

closes odoo/upgrade#6148

Related: odoo/odoo#168666
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>