[FIX] maintenance/saas~16.5: set default for newly introduced fields
Cause: In version 17.0, new fields (repeat_unit, repeat_type, repeat_until) were
introduced in the maintenance module. These fields are set as required in
the view 'maintenance.hr_equipment_request_view_form', but they are only
visible if the request is a recurrent maintenance request, which causes a
validation error when changing the status.

Reason: default not set for the mentioned fields. in source
[repo-odoo/odoo](https://github.com/odoo/odoo/blob/d7cfef9c51461a595f3f46c7b91e5c56241d4af3/addons/maintenance/models/maintenance.py#L265)
the default was set but not in the migration script
[ref](https://github.com/odoo/upgrade/blob/29dc34ef47b469d29ea103cc8877f83b5fbadad4/migrations/maintenance/saas~16.5.1.0/pre-migrate.py#L7)

Solution: set the default for the newly introduced fields.

Note: This is not a bug in the general version of 17.0; it is introduced
after an upgrade from version 16.0 to 17.0.

closes odoo/upgrade#6147

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>