[FIX] repair/saas~16.4: recorrect fix to save all operations
There was an 'Operations' tab in 'repair_order' which refers to the 'repair_fee'
field. This page allowed to link service-type products to repairs. While
upgrading from version > saas~16.3 this tab has been removed. With the upgrade
fix in version saas~16.4, the data of 'operations' of repair_order has been
saved as 'internal notes'. While upgrading to V17.0, in customer's database,
there were still some operations that were not migrated due to the fact that
previously the relation was linked with the repair_fee_line_tax and repair_order
table which should not be the case.

As a matter of fact, the paired should be between the repair_order, repair_fee
, and repair_fee_line_tax table. After this fix, it will save all the operations
from the repair order and save them as internal notes.

opw-3943391,3864870
upg-1728608,1493549

closes odoo/upgrade#6151

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>