[FIX] helpdesk/saas~15.4: Also update mt_ticket_new
The record `mt_team_ticket_new` depends on the record `mt_ticket_new`.
If the latter is missing, when the former gets updated it will trigger
an error.

By adding it to the list of records to update, we ensure it exists.

upg-1731550

```
ValueError: External ID not found in the system: helpdesk.mt_ticket_new
[...]
odoo.tools.convert.ParseError: while parsing None:78, somewhere inside
<record id="mt_team_ticket_new" model="mail.message.subtype">
	<field name="name">Ticket Created</field>
	<field name="sequence">0</field>
	<field name="res_model">helpdesk.team</field>
	<field name="default" eval="True"/>
	<field name="parent_id" ref="mt_ticket_new"/>
	<field name="relation_field">team_id</field>
</record>
```

closes odoo/upgrade#6154

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>