[IMP] account_peppol/saas~17.4: refactor partner check
Currently, verifying whether a partner exists on Peppol network
and whether they can receive a particular document type is
unnecessarily complex.
We have three fields for the same purpose,
even though we could use just one.

Additionally, partners are shared between companies.
Each company in the database can be registered on Peppol separately
in different edi modes: demo, test, prod.

Each edi mode has its own partner verification check, so the
verification state needs to be company-dependent.
At the same time, it needs to be computed on the fly
when a user updates peppol_eas, peppol_endpoint or ubl_cii_format
on that partner.
In that moment, we recompute the verification state for every
company in the db.

See also:
odoo/odoo#169650

closes odoo/upgrade#6164

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>