[IMP] mrp/saas~17.4: remove confirm_cancel field
After merging the community PR, Odoo will prompt for confirmation when
the cancel button on a transfer or MO is pressed. This will help
prevent accidental cancellations. Additionally, the confirmation warning
should always be present for both cancel buttons on the MO to ensure
consistency. The `confirm_cancel` field logic was removed. In previous
versions, the field was kept, but was no longer used. In master, the
field should be removed.

closes odoo/upgrade#6167

Related: odoo/odoo#165841
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>