[FIX] sale_subscription/saas~15.3: fix sol constraint violation
Some clients have sale **subscription** lines with NULL `product_id`.
This is not valid in standard, in 15:
https://github.com/odoo/enterprise/blob/35e8b03955c0f7536c13d4ac9938ffa739bc7938/sale_subscription/models/sale_subscription.py#L1262-L1264

We still provide this patch because we have observed that it's common
for clients to install custom modules that mimic the behavior from sale
**order** lines: add a `display_type` that allows to have null
`product_id` when the display type is not NULL.

In such situations the upgrade fail with constraint violation error:
```
  File "/tmp/tmp60xgcdaw/migrations/sale_subscription/saas~15.3.1.1/pre-migrate.py", line 409, in migrate
    cr.execute(
  File "/home/odoo/src/odoo/16.0/odoo/sql_db.py", line 321, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.CheckViolation: new row for relation "sale_order_line" violates check constraint "sale_order_line_accountable_required_fields"
```
Here are the standard constraints in 16:
https://github.com/odoo/odoo/blob/30b35bd3c9129ed39cc2f51ece4a2850d4a898f6/addons/sale/models/sale_order_line.py#L22-L29

To avoid the issue we upgrade the order lines as notes. If clients want
to recover `line_section` then specific scripts can be made. To be safe
we only set as notes lines that _also_ have zero qty --besides null
product. If there are lines with null product and non zero quantities
the script will still fail.

closes odoo/upgrade#6174

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>