[FIX] loyalty/0.0.0 : Loyalty program performance issue
In any upgraded database of version 16 or above, While creating a new record or
opening the existing record in the 'Gift cards & e-wallet' menu,it's taking too
much time due to field 'all_discount_product_ids' taking time in computation.So
for that reason, while we open the record,it will compute display_name of all
the products.In many case, rewards are available on all products,causing the
computation of the m2m to fill it with the ids of all products.This caused
performance issues on all DBs with lots of products (~5k).

As such, new databases don't suffer from this performance issue, an
ir_config_parameter has been introduced to opt into the new behaviour.
But there are many existing database which have a lot of product and performance
issue related to this.

This solution ensures that existing databases with a large number of products
will have the loyalty.compute_all_discount_product_ids parameter set to False,
thereby preventing performance issues when creating or opening records in the
'Gift cards & e-wallet' menu.

opw- 3849879,3876884,3745781,3500153
upg - 1570770,1483594,1382308

closes odoo/upgrade#6181

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>