[FIX] sale_subscription/saas~15.3.1.1: handle custom field restoring
In this #6132 during handling moving translations of custom manual fields from
`sale.subscription*` to `sale.order*` we moved restoring of data on
manual fields to `pre-migrate` from `end-migrate`, because we should
execute the restoring before module loading for translatable fields.

But we have introduced another issue without noticing it at that time.
Right now when running `pre` script this actions are executed in this order:

1. Change the referencing model to `sale.order.*` from `sale.subscription.*`
   on **ir.model.fields**  (pre-script)
2. Create new column with same name and type on `sale_order*` (pre-script)
3. Restore the data of that column from table `sale_subscription_*` to
   `sale_order_*`. (pre-script)
4. Switch the pointing foreign key from `sale_subscription*` to `sale_order*`.
   Basically if there is record pointing to `sale_subscription*`,this step
   re-points that record to `sale_order*`. Apparently the value in pointing
   record will be changed (pre-script)

The problem here is, we need to execute the 4th step before 3rd.
Because if there is custom manual related field between `sale_subscription*`
tables, for example if there is custom related field on `sale.subsctiption.line`
which is pointing to `sale.subscription.template`. The value on the pointing
field should be changed before restoring the fields.

For fixing the issue I have moved the restoring to the end of `pre` script.

Steps to reproduce:
1. Install `sale_subscription` on `15.0`
2. Add related field on `sale.subscription.line` which points to
   `sale.subscription.template`
3. By default you will have 2 records in `sale.subscription.template` and
   1 record in `sale.order.template`. Point all of your subscription lines
   to the sale template `Yearly Subscription`
   (i.e `sale_subscription.yearly_subscription`)
4. Upgrade to `16.0`

Then look at the table `sale_subscription_lines` before upgrade and
look `sale_order_line` after upgrade.

Before upgrade all subscription lines are pointing to `Yearly Subscription`:

```SQL
> SELECT ssl.id, ssl.name, sst.name
    FROM sale_subscription_line ssl
    JOIN sale_subscription_template sst
      ON ssl.x_studio_many2one_field_awk5p = sst.id
ORDER BY ssl.id;

 id |                     name                      |        name
----+-----------------------------------------------+---------------------
  1 | Office Cleaning Service                       | Yearly Subscription
  2 | Office Cleaning Service (yearly subscription) | Yearly Subscription
  3 | Office Cleaning Service                       | Yearly Subscription

> SELECT id, name FROM sale_subscription_template WHERE id = 2;
 id |         name
----+--------------------
  2 | Yearly Subscription

```

After the upgrade they are pointing to `Upgraded Monthly Subscription` because
we restored the field `x_studio_many2one_field_awk5p` before we switch the
foriegn keys to the model `sale.subscription.template`.
In other words the id of `Yearly Subscription` template was 2 in
`sale.subscription.template` but it's id is not not 2 in `sale.order.template`

```SQL
> SELECT id, name, new_sale_order_line_id
    FROM sale_subscription_line;

 id |                     name                      | new_sale_order_line_id
----+-----------------------------------------------+------------------------
  1 | Office Cleaning Service                       |                     45
  3 | Office Cleaning Service                       |                     46
  2 | Office Cleaning Service (yearly subscription) |                     47

> SELECT sol.id, sol.name, sot.name
    FROM sale_order_line sol
    JOIN sale_order_template sot
      ON sol.x_studio_many2one_field_awk5p = sot.id
ORDER BY sol.id;

 id |                     name                      |             name
----+-----------------------------------------------+-------------------------------
 45 | Office Cleaning Service                       | Upgraded Monthly Subscription
 46 | Office Cleaning Service                       | Upgraded Monthly Subscription
 47 | Office Cleaning Service (yearly subscription) | Upgraded Monthly Subscription

> SELECT id, name FROM sale_order_template WHERE id = 2;

 id |             name
----+-------------------------------
  2 | Upgraded Monthly Subscription

```

closes odoo/upgrade#6175

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>