[FIX] base/0.0.0: cleanup empty default
When default value is unset in web_studio, it calls the route[^1]
with empty string which is causing setting empty string as default on
`ir.default` model.

On the other hand, when new record is being created, in `default_get`[^2]
method tries to get the default from `ir.default` first then it looks
to the actual default of the field.

The problem is occuring on selection, required fields like this:

1. The method `convert_to_cache`[^3] of selection fields is returning **None**
   if the value is empty string.
2. So empty string would be converted to `None` here[^4]
3. That None will be converted to False here[^5] because
   it will call[^6] at the end
4. Then when it comes to create the record, value of that column
   becomes `None` here [^7] because at the end it will come here[^8]
   with value `None` or `False` and value of that column will be `None`

At the end, this[^9] will cause Not NULL violation on that column when you try
to add record to the model which that required selection field is pointing at.

For reproducing the issue:

1. Install stock in `16.0`
2. Go to developer mode and install studio
3. Set default value on field `tracking` on `product.template`
4. Then unset that default by just setting empty line.
5. Try to create `product.template` record without setting tracking.

You will not be able to create it. Also during upgrade if it tries to add
new or missing records on that model, it will cause NOT NULL violation error.

As we see here, there is a bug on studio, it needs to remove the overridden
default when user selects empty line while setting default. We need to introduce
new endpoint for removing existing default on `web_studio` module and on studio
code there needs to be adaption with that new route and already existing one
instead of calling old route with empty string in certain cases.

On the other hand having empty default on required selection fields is wrong.
So in this PR we remove all overridden defaults of required, selection fields.

TBG-1131

[^1]: https://github.com/odoo/enterprise/blob/04586fcc2fa7e716682803918e3103d14df9c2de/web_studio/controllers/main.py#L1667
[^2]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/models.py#L1487-L1497
[^3]: https://github.com/odoo/odoo/blob/93c8e362e397b89ab28eea9f6791a06d051a8111/odoo/fields.py#L2833
[^4]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/models.py#L1515
[^5]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/models.py#L1516
[^6]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/fields.py#L942
[^7]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/models.py#L4775
[^8]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/fields.py#L907
[^9]: https://github.com/odoo/odoo/blob/a889fc5be2d85b659749bae103b241ffa045e6f8/odoo/models.py#L4796-L4801

closes odoo/upgrade#6184

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>