[FIX] base/0.0.0: remove failing saas view
During upgrade of saas databases, this view from saas_trial is disabled
because of a missing field. As the module is not available during upgrade,
the field is not available, and incorrectly reported as missing.

As the module will be update right after restore on the saas, removing
the view is not an issue.

closes odoo/upgrade#6189

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>