[FIX] repair/saas~16.4: add check for move if unlinked
As we perform `stock.move` operations on newly created stock moves from
repair order lines, when actions are performed to reserve and confirm those
moves, in action_explode
[here](https://github.com/odoo/odoo/blob/41fcd23772a84e6250865ec299479a0058425102/addons/mrp/models/stock_move.py#L431-L464)
if any product in the moves has a BOM defined,
we create stock moves for the BOM lines of the respective product's BOM and
[unlink](https://github.com/odoo/odoo/blob/41fcd23772a84e6250865ec299479a0058425102/addons/mrp/models/stock_move.py#L464)
the main stock move. After the main move is unlinked, the iteration
still includes that unlinked `stock.move`, which will block the migration with
the following MissingError:
```py
2024-06-27 10:01:26,659 27 INFO apan_1775369_16.0_17.0 odoo.models.unlink: User #1 deleted stock.move records with IDs: [77977]

File "/tmp/tmps0idklex/migrations/repair/saas~16.4.1.0/post-migrate.py", line 26, in migrate
    move._trigger_scheduler()
  File "/home/odoo/src/odoo/17.0/addons/stock/models/stock_move.py", line 2115, in _trigger_scheduler
    ('product_id', '=', move.product_id.id),
  File "/home/odoo/src/odoo/17.0/odoo/fields.py", line 2907, in __get__
    return super().__get__(records, owner)
  File "/home/odoo/src/odoo/17.0/odoo/fields.py", line 1188, in __get__
    raise MissingError("\n".join([
odoo.exceptions.MissingError: Record does not exist or has been deleted.
(Record: stock.move(77977,), User: 1)
```
To avoid this MissingError, As it is already done
[here](https://github.com/odoo/odoo/blob/41fcd23772a84e6250865ec299479a0058425102/addons/repair/models/stock_move.py#L79-L82)
we have updated the move value if the any new moves are created
and main move got unlinked.

UPG - 1775369, 1776553
OPW - 4010990, 4012601

closes odoo/upgrade#6195

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>