[IMP] documents/saas~17.5: share-pocalyse 🔥
Purpose
=======
The documents application has been refactored (the access rules changed, the shares have
been removed, etc).

This commit migrates the changes.

Folder
======
Before, we had a separate model for the folder of the document app. Now, everything is
stored on the same model, <documents.document>.

Facet / tags
============
Before, we had <documents.facet> that were linked to a <documents.folder>, and the tags
were linked to those <documents.facet>. Now, there's no more <documents.facet>, and the
tags can be added in any folder.

We remove duplicated tags (if 2 tags had the same name, but different facets).

Access
======
Before, the accesses were on <documents.document> and on <documents.share> at the same time.
- <documents.document> stores the groups that are allowed to read / write
- <documents.share> stores `allow_upload` for the people that received a token

Now, there are 2 important fields on <documents.document>
- access_link (view, edit or none) describe what we can do with a token
- access_internal (view, edit or none) describe what an internal user can do

To have more flexibility, we also have a new model, <documents.access> for which we can
customize the access for each user. That model has a `role` field, and if the role is
Falsy, we read `access_link` (it's used to know when a user clicked on a link with a token).

Share
=====
Before, we had a model for sharing documents. We could share one folder, one domain,
a list of documents, etc. Now, everything is stored on the document. So, we have only
one access token per document (file, or "old folder").

But, we want the old link to still work (at least that the users can still download
the documents we sent them). To solve that problem, we create a temporary model
<documents.redirect>, to translate the token to the only one we kept on
<documents.document>.

There's one problem tough, some share can be "writable" and other not. And we don't want
to give write access to the people that received a "read only" link. For simplicity, we
switch all share to readonly mode, and we will need to re-enable the write access if
needed (the most important thing is that people should be able to download their documents).

Some share could have `include_sub_folders`, which is not easily possible with the new
implementation. It has been decided that it was ok to loose that type of share.

Before, shares could have an expiration date, but we don't support that feature anymore,
and we don't want to extend the share for an infinite long time, so we decided to not
migrate them.

Alias
=====
Before, the aliases were stored on <documents.share>. And so, we could have many alias for
the same folder. Now, the alias are configured on the document. And so, from the documents
interface, it's not possible to have many alias for the same folder (but it's still
possible to do it, from the technical form view if <mail.alias>).

Some tags can be added automatically on the document when it's sent by email, those tags
are not linked to <documents.document> (instead of <documents.share>).

Again, we want to loose as less information as possible, and so we want all aliases
to still work after the migration. So, for each document, we choose one alias to link to
<documents.document> (in priority the alias with tag), and we move the tag to
`alias_tag_ids`. For other aliases, we keep them hidden, and their tag stay in
`alias_defaults`.

User specific
=============
Before, a folder could have been marked as `user_specific` or `user_specific_write`.
When that flag is True, a user can see / edit only the documents for which he's the owner.

That behavior can be simulated with <documents.access>. We also need to force
`is_access_link_hidden`, so other users won't be able to see the document if we
decide to share it (and change `access_link`). We also need to remove the owner,
otherwise he will be able to write on the document.

For user_specific_write, we can just keep the owner on the document,
and he will have a write access.

Company
=======
Before, the company was set on the folder, now, the company is set on the
<documents.document>, so on the files and on the folders.

So, we propagate the company, from the top to the bottom.

Activity
========
In the migration only one share is migrated to a document folder. The other
are migrated as simple alias which we name secondary alias.

The share (and now the document folder) supports the creation of an activity
when an attachment is received by email through an alias. So for the main alias
we just copy the activity fields to the folder that replace the share/alias.

In order to keep that feature for the secondary aliases, we set in the
alias_defaults the values of the activity creation fields of the share
(activity_option, ..., prefixed with "create_") so that the document "template"
created when receiving a mail through the alias has the information to create
the activity on the documents created from all the attachments of the email
(using the templates). Indeed, that first created document (template) is
archived right away and only serve as a template to create the real documents.
So we can use the create_xx of that document to create the activities on the
final document holding the attachment of the email.

To sum up, we migrate the main share activity create fields to the folder that
replace that share and we migrate the activity create fields of the secondary
share in the alias_defaults of the corresponding secondary alias.

Activity fields (activity_option, activity_type_id, activity_summary, ...) of
document share were defining the activity to create when uploading a document
to the share. Now that document acts as a share, we have moved those field to
it. But as those fields were conflicting with mail.activity.mixin fields that
document inherits, we have renamed with a prefix create_.

To be consistent and not have a special case for the method
documents_set_activity of document, we also have renamed the activity fields of
workflow_rule that we migrate here (including view and data).

Finally, we also update the request templates (for example access_link was
changed to acces_url).

Task-4014621
COM odoo/odoo/pull/180264
ENT odoo/enterprise/pull/65655

Part-of: odoo/upgrade#6226
Related: odoo/odoo#180264
Related: odoo/enterprise#65655
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Pierre-Yves Dufays <pydu@odoo.com>