[IMP] delivery/saas~16.2: avoid failure when installing module
In saas~16.2 if we force install module 'delivery' as a new dependency
for 'website_sale' we currently have a uom consistency check as an early
sign that the installation will fail. But we dont really need it because
we can create the stored computed field like we did when the field was
introduced in saas~16.1 and this will also avoid memory errors on compute
if the sale order line table it too big, which is common.

Here we fix another oversight, the field 'shipping_weight' was never
moved from 'delivery' to 'stock_delivery'.

closes odoo/upgrade#6205

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>