[FIX] account/16.0: preserve old tax report behaviour
opw-3820695

Before Odoo 16, lines with no `tag_name` and no `formula` evaluated to the sum
of the their children lines with a `tag_name` set (ignoring, for instance those
with a `formula` set, instead). This behaviour was not preserved and is not
directly achieved by `sum_children` or similar shortcuts. Therefore it is
achieved by computing the full blown aggregation formula, before actually
migrating those lines.

closes odoo/upgrade#6217

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>