[IMP] l10n_latam_check/saas~17.4.1.0: from one to multiple checks
Before the checks where payments with extra fields on the payments
and a one2many between payments as in some countries, an original
check received from a customer can be used to pay a supplier.

The problem however is that they usually pay with a lot of checks
at the same time and for this to work, we decided to have a separate
check objects and allow multiple checks per payment.

Community PR: https://github.com/odoo/odoo/pull/169887

closes odoo/upgrade#6233

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>