[IMP] point_of_sale/saas~17.4: Display scenario when no result is found
Before when the user searches for a config and no result is found, the
user saw an empty screen.

Now, the user will see a message that no result was found and different
buttons with options to create a new config or to create new products.

Community PR: https://github.com/odoo/odoo/pull/172480

closes odoo/upgrade#6238

Taskid: 4035349
Related: odoo/odoo#172480
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>