[IMP] account_reports/saas~17.4: Report Footnotes become annotations
The aim of this commit is renaming the footnotes model for
annotations once.

The new model authorizes the user to add a date to the annotation
to show it only for a specific period. It also allows to link it
to a fiscal position. For the migration, we decide to keep existing
foot notes "as it" and we don't add any date or fiscal position.

See odoo/enterprise#48530

task-3301125

closes odoo/upgrade#6244

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>