[ADD] stock_fleet/saas~17.4: move field for common usage
We add a new module for dispatch management. However it's base on some
constraints with the weight and the volume. So we moved them from
stock_delivery to stock.

We also made some cleaning in stock_batch_picking with the xml_id
and removed a useless field in stock_delivery

OC odoo/odoo#172309
OE odoo/enterprise#66259

TASK - 3729151

closes odoo/upgrade#6247

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>