[FIX] account/17.0.1.2: fix translations of view
In V17.0,  the <br/> tag has been added in the report_invoice_document view
the <br/> tag is missing in translations for other languages, except (en_US)
However,during the upgrade, the translation for this was not updated as per
the standard, the translation from the previous version is taken.

This PR will manually update translations of the view.

closes odoo/upgrade#6250

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>